hack(match): REVERT ME LATER fix type errors
When bl-model changed (which has already happened in the API) these
pages broke because they relied on the old schema. This commit
temporarily fixes the TypeScript errors by setting the Match type to any
in those files. This commit will be tagged with
'revert-me-bl-model-upgrade'; when it is time to fix these pages, revert
this commit and fix them.