fix(MatchDetail): fix 5s load time
Using the same key 'matches/me' for useSWR in MatchesList and
MatchDetail caused conflict because they used different fetchers
(MatchDetail extracted the specific relevant match), which when cached
caused both pages to misbehave on navigation between them. This first
manifested in errors when going detail -> list, so I first changed it to
a full-refresh using the MuiLink instead of DynamicLink. This had the
result of purging the cache, fixing the symptom in one direction but
hurting performance. Then, MatchDetail took 5s to load because it first
used the cached version from MatchesList, until the refreshInterval
elapsed. It was fixed by using same fetcher for both pages, and
extracting the match outside the fetcher. Another solution could be to
use different keys, but this way the cache could be properly utilized,
potentially preventing refetch on navigation.