fix(signup): remove ref warning
Controller from react-hook-form cannot take a ref, so that had to be
excluded. In the docs, they don't pass the register()-return to
Controller at all, but when I tried their way I didn't get validation,
so this was the simplest fix.

Well partially simplest, I also prevented the invalid-look of the date
input when only a partial date had been inputted, but the field had not
blurred.